requirements.txt
boto3
pandas
streamlit
python-dotenv
matplotlib
scikit-learn
